Fixes the native application a bit more
